PK값 Tsid 포맷 직접 할당 전략으로 변경 (#214)

+++ migrations/notice-post/1706442705381-notice-post.ts
@@ -15,7 +15,7 @@ export class NoticePost1706442705381 implements MigrationInterface {
           generatePrimaryColumn('공지 게시글 고유 ID'),
           {
             name: 'user_id',
-            type: 'int',
+            type: 'bigint',
             unsigned: true,
             isNullable: false,
             comment: '게시글 작성 유저 고유 ID',
